SDL: Fix wrong size after leaving fullscreen with SDL2
In SDL2, calling SDL_DestroyWindow with a fullscreen window that
doesn't use SDL_WINDOW_FULLSCREEN_DESKTOP causes a resize
event to be sent with the old size. When leaving fullscreen that
event was processed after recreating the window and would cause
the window resolution to be changed back to the fullscreen
resolution.

This commits add an explicit call to SDL_SetWindowSize after
SDL_CreateWindow which cases another resize event to be sent.
This way when we destroy and recreate a window the last resize
event to be processed has the correct size. Also SDL_SetWindowSize
sends a SDL_WINDOWEVENT_SIZE_CHANGED event and not
SDL_WINDOWEVENT_RESIZED. So we need to call handleResizeEvent()
for both in order for the fix to work.