BUILD: Respect all build tool overrides from environment
Previously, only CXX could be overridden by the environment, which
made it impossible to correctly set tools for cross-compiler
toolchains which were not compatible with the default system tools.

ld still always uses CXX because an LD override in the environment
is not the correct way to link a C++ program in any case.