DINGUX: Remove dead code
When compiled with C++11, it was revealed that initSize has the
wrong method signature so has not been overriding initSize
and so has not been doing anything. Apparently this is OK, since
nobody has complained about it, so just get rid of it.

The setGraphicsMode(const char *) signature was also wrong, and
has no implementation code anyway.